+++ src/org/wildstang/yearly/auto/programs/ShootGear.java
@@ -15,7 +15,7 @@ import org.wildstang.yearly.auto.steps.WaitStep;
 public class ShootGear extends AutoProgram
 {
 
-   private double delayWhileShooting;
+   private long delayWhileShooting;
 
    @Override
    public void initialize()
